+++ src/view/app/ModuleDetails/profilerModuleDetails.tsx
@@ -1,35 +1,38 @@
 import * as React from "react";
 import { useState, useMemo } from "react";
 import {
-  PresentationData,
   ModuleDetails,
   CalledModules,
   LineSummary,
+  PresentationData,
 } from "../../../common/PresentationData";
 import DataGrid from "react-data-grid";
-import type { Column, SortColumn } from "react-data-grid";
+import type { Column, FormatterProps, SortColumn } from "react-data-grid";
 import * as columnDefinition from "./column.json";
 import "./profilerModuleDetails.css";
+import ModuleDetailsTable from "./components/ModuleDetailsTable";
+import { getVSCodeAPI } from "../utils/vscode";
+import PercentageFill from "./components/PercentageFill";
+import { Box } from "@mui/material";
+import ModuleDetailsSettings from "./components/ModuleDetailsSettings";
+import { useModuleDetailsSettingsContext } from "./components/ModuleDetailsSettingsContext";
+import { OpenFileTypeEnum } from "../../../common/openFile";
 
-interface ModuleColumn extends Column<any> {
+interface ProfilerModuleDetailsProps {
+  presentationData: PresentationData;
+  moduleName: string;
+}
+
+interface GenericModuleColumn extends Column<any> {
   key: string;
   name: string;
   width: string;
 }
 
-interface IConfigProps {
-  vscode: any;
-  presentationData: PresentationData;
-  selectedRow: any;
-  onRowSelect: (row: any) => void;
-  moduleName: string;
-}
-
-const filterCSS: React.CSSProperties = {
-  inlineSize: "100%",
-  padding: "4px",
-  fontSize: "14px",
-};
+interface ModuleColumn extends GenericModuleColumn {}
+interface CallingColumn extends GenericModuleColumn {}
+interface CalledColumn extends GenericModuleColumn {}
+interface LineColumn extends GenericModuleColumn {}
 
 const defaultModuleSort: SortColumn = {
   columnKey: "totalTime", // Sort by the "totalTime" column by default
@@ -41,31 +44,38 @@ const defaultLineSort: SortColumn = {
   direction: "ASC", // Use ascending order by default
 };
 
-function moduleRowKeyGetter(row: ModuleDetails) {
-  return row.moduleID;
-}
-
 const addConditionalFormatting = (
-  columns: Array<ModuleColumn>
-): Array<ModuleColumn> => {
+  columns: Array<GenericModuleColumn>
+): Array<GenericModuleColumn> => {
+  const addPercentageFormat = (value: number) => (
+    <PercentageFill value={value} />
+  );
+
+  const addLinkFormat = (row: ModuleDetails | LineSummary, key: string) => (
+    <Box className={row.hasLink ? "link-cell" : ""}>{row[key]}</Box>
+  );
+
   return columns.map((column) => {
     if (column.key === "moduleName" || column.key === "lineNumber") {
       return {
         ...column,
-        formatter: ({ row }: { row: ModuleDetails | LineSummary }) => {
-          const className = row.hasLink ? "link-cell" : "";
-          return <div className={className}>{row[column.key]}</div>;
-        },
+        formatter: (props: FormatterProps<ModuleDetails | LineSummary>) =>
+          addLinkFormat(props.row, column.key),
+      };
+    }
+    if (
+      column.key === "calleePcntOfSession" ||
+      column.key === "callerPcntOfSession"
+    ) {
+      return {
+        ...column,
+        formatter: (props: FormatterProps<CalledModules>) =>
+          addPercentageFormat(props.row[column.key]),
       };
     }
     return column;
   });
 };
-
-type ModuleComparator = (a: ModuleDetails, b: ModuleDetails) => number;
-type CallingComparator = (a: CalledModules, b: CalledModules) => number;
-type CalledComparator = (a: CalledModules, b: CalledModules) => number;
-type LineComparator = (a: LineSummary, b: LineSummary) => number;
 
 function getComparator(sortColumn: string) {
   switch (sortColumn) {
@@ -96,120 +106,99 @@ function getComparator(sortColumn: string) {
   }
 }
 
-function getModuleComparator(sortColumn: string): ModuleComparator {
-  return getComparator(sortColumn);
-}
-
-function getCallingComparator(sortColumn: string): CallingComparator {
-  return getComparator(sortColumn);
-}
-
-function getCalledComparator(sortColumn: string): CalledComparator {
-  return getComparator(sortColumn);
-}
-
-function getLineComparator(sortColumn: string): LineComparator {
-  return getComparator(sortColumn);
-}
-
-function ProfilerModuleDetails({
+const ProfilerModuleDetails: React.FC<ProfilerModuleDetailsProps> = ({
   presentationData,
-  vscode,
-  selectedRow,
-  onRowSelect,
   moduleName,
-}: IConfigProps) {
-  const [moduleRows, setModuleRows] = useState(presentationData.moduleDetails);
+}) => {
+  const [moduleRows, setModuleRows] = useState<ModuleDetails[]>(
+    presentationData.moduleDetails
+  );
   const [selectedModuleRow, setSelectedModuleRow] =
     useState<ModuleDetails | null>(null);
   const [sortModuleColumns, setSortModuleColumns] = useState<
     readonly SortColumn[]
   >([defaultModuleSort]);
-  const [filteredModuleRows, setFilteredModuleRows] = useState(moduleRows);
 
-  const [callingRows, setCallingRows] = useState(
-    presentationData.calledModules
-  );
-  const [selectedCallingRows, setSelectedCallingRows] = useState(
-    presentationData.calledModules
-  );
+  const [selectedRow, setSelectedRow] = useState<ModuleDetails>(null);
+  const [selectedCallingRows, setSelectedCallingRows] = useState<
+    CalledModules[]
+  >(presentationData.calledModules);
   const [sortCallingColumns, setSortCallingColumns] = useState<
     readonly SortColumn[]
   >([]);
 
-  const [calledRows, setCalledRows] = useState(presentationData.calledModules);
-  const [selectedCalledRows, setSelectedCalledRows] = useState(
+  const [selectedCalledRows, setSelectedCalledRows] = useState<CalledModules[]>(
     presentationData.calledModules
   );
   const [sortCalledColumns, setSortCalledColumns] = useState<
     readonly SortColumn[]
   >([]);
 
-  const [lineRows, setLineRows] = useState(presentationData.lineSummary);
-  const [selectedLineRows, setSelectedLineRows] = useState(
+  const [selectedLineRows, setSelectedLineRows] = useState<LineSummary[]>(
     presentationData.lineSummary
   );
   const [sortLineColumns, setSortLineColumns] = useState<readonly SortColumn[]>(
     [defaultLineSort]
   );
 
-  const [formattedModuleColumns] = useState(
-    addConditionalFormatting(columnDefinition.moduleColumns)
+  const [moduleNameFilter, setModuleNameFilter] =
+    React.useState<string>(moduleName);
+
+  const vscode = getVSCodeAPI();
+
+  const formattedModuleColumns: ModuleColumn[] = addConditionalFormatting(
+    columnDefinition.moduleColumns
+  );
+  const callingColumns: CallingColumn[] = addConditionalFormatting(
+    columnDefinition.CallingColumns
+  );
+  const calledColumns: CalledColumn[] = addConditionalFormatting(
+    columnDefinition.CalledColumns
+  );
+  const formattedLineColumns: LineColumn[] = addConditionalFormatting(
+    columnDefinition.LineColumns
   );
 
-  const [formattedLineColumns] = useState(
-    addConditionalFormatting(columnDefinition.LineColumns)
-  );
-
-  const calledColumns = columnDefinition.CalledColumns;
-  const callingColumns = columnDefinition.CallingColumns;
+  const settingsContext = useModuleDetailsSettingsContext();
 
   const sumTotalTime = presentationData.moduleDetails.reduce(
     (acc, module) => acc + module.totalTime,
     0
   );
 
-  const [filters, _setFilters] = React.useState({
-    columns: {},
-    enabled: true,
-  });
-  const filtersRef = React.useRef(filters);
-  const setFilters = (data) => {
-    filtersRef.current = data;
-    _setFilters(data);
+  const filterTables = (row: ModuleDetails) => {
+    if (!row) {
+      return;
+    }
+
+    setSelectedCallingRows(
+      presentationData.calledModules.filter(
+        (element) => element.calleeID === row.moduleID
+      )
+    );
+    setSelectedCalledRows(
+      presentationData.calledModules.filter(
+        (element) => element.callerID === row.moduleID
+      )
+    );
+    setSelectedLineRows(
+      presentationData.lineSummary.filter(
+        (element) => element.moduleID === row.moduleID
+      )
+    );
   };
 
-  React.useEffect(() => {
-    filterByModuleName(moduleName);
-  }, [moduleName]);
-
-  calledColumns.forEach((column) => {
-    if (column.key === "calleePcntOfSession") {
-      addPercentage(column);
-    }
-  });
-
-  callingColumns.forEach((column) => {
-    if (column.key === "callerPcntOfSession") {
-      addPercentage(column);
-    }
-  });
-
-  const filterByModuleName = (moduleName: string) => {
-    if (moduleName !== "") {
-      filterColumn("moduleName", moduleName);
-      setSelectedModuleRow(null);
-    }
-  };
-
-  const sortedModuleRows = useMemo((): readonly ModuleDetails[] => {
-    if (sortModuleColumns.length === 0) {
-      return filteredModuleRows;
+  const getSortedRows = (
+    columns: readonly SortColumn[],
+    rows: ModuleDetails[] | CalledModules[] | LineSummary[]
+  ) => {
+    if (columns.length === 0) {
+      return rows;
     }
 
-    const sortedRows = [...filteredModuleRows].sort((a, b) => {
-      for (const sort of sortModuleColumns) {
-        const comparator = getModuleComparator(sort.columnKey);
+    return [...rows].sort((a, b) => {
+      for (const sort of columns) {
+        const comparator = getComparator(sort.columnKey);
         const compResult = comparator(a, b);
         if (compResult !== 0) {
           return sort.direction === "ASC" ? compResult : -compResult;
@@ -217,6 +206,13 @@ function ProfilerModuleDetails({
       }
       return 0;
     });
+  };
+
+  const sortedModuleRows = useMemo((): readonly ModuleDetails[] => {
+    const sortedRows = getSortedRows(
+      sortModuleColumns,
+      moduleRows
+    ) as ModuleDetails[];
 
     if (sortedRows.length > 0 && selectedModuleRow === null) {
       setSelectedModuleRow(sortedRows[0]);
@@ -224,242 +220,38 @@ function ProfilerModuleDetails({
     }
 
     return sortedRows;
-  }, [filteredModuleRows, sortModuleColumns]);
+  }, [moduleRows, sortModuleColumns]);
 
   const sortedCallingRows = useMemo((): readonly CalledModules[] => {
-    if (sortCallingColumns.length === 0) {
-      return selectedCallingRows;
-    }
-
-    return [...selectedCallingRows].sort((a, b) => {
-      for (const sort of sortCallingColumns) {
-        const comparator = getCallingComparator(sort.columnKey);
-        const compResult = comparator(a, b);
-        if (compResult !== 0) {
-          return sort.direction === "ASC" ? compResult : -compResult;
-        }
-      }
-      return 0;
-    });
+    return getSortedRows(
+      sortCallingColumns,
+      selectedCallingRows
+    ) as CalledModules[];
   }, [selectedCallingRows, sortCallingColumns]);
 
   const sortedCalledRows = useMemo((): readonly CalledModules[] => {
-    if (sortCalledColumns.length === 0) {
-      return selectedCalledRows;
-    }
-
-    return [...selectedCalledRows].sort((a, b) => {
-      for (const sort of sortCalledColumns) {
-        const comparator = getCalledComparator(sort.columnKey);
-        const compResult = comparator(a, b);
-        if (compResult !== 0) {
-          return sort.direction === "ASC" ? compResult : -compResult;
-        }
-      }
-      return 0;
-    });
+    return getSortedRows(
+      sortCalledColumns,
+      selectedCalledRows
+    ) as CalledModules[];
   }, [selectedCalledRows, sortCalledColumns]);
 
   const sortedLineRows = useMemo((): readonly LineSummary[] => {
-    if (sortLineColumns.length === 0) {
-      return selectedLineRows;
-    }
-
-    return [...selectedLineRows].sort((a, b) => {
-      for (const sort of sortLineColumns) {
-        const comparator = getLineComparator(sort.columnKey);
-        const compResult = comparator(a, b);
-        if (compResult !== 0) {
-          return sort.direction === "ASC" ? compResult : -compResult;
-        }
-      }
-      return 0;
-    });
+    return getSortedRows(sortLineColumns, selectedLineRows) as LineSummary[];
   }, [selectedLineRows, sortLineColumns]);
-
-  formattedModuleColumns.forEach((column) => {
-    if (column.key === "moduleName") {
-      column["headerRenderer"] = function ({
-        onSort,
-        sortDirection,
-        priority,
-      }) {
-        function handleClick(event) {
-          onSort(event.ctrlKey || event.metaKey);
-        }
-
-        function handleInputKeyDown(event) {
-          filterColumn(column.key, event.target.value);
-        }
-
-        return (
-          <React.Fragment>
-            <div className={filters.enabled ? "filter-cell" : undefined}>
-              <span
-                tabIndex={-1}
-                style={{
-                  cursor: "pointer",
-                  display: "flex",
-                }}
-                className="rdg-header-sort-cell"
-                onClick={handleClick}
-              >
-                <span
-                  className="rdg-header-sort-name"
-                  style={{
-                    flexGrow: "1",
-                    overflow: "clip",
-                    textOverflow: "ellipsis",
-                  }}
-                >
-                  {column.name}
-                </span>
-                <span>
-                  <svg
-                    viewBox="0 0 12 8"
-                    width="12"
-                    height="8"
-                    className="rdg-sort-arrow"
-                    style={{
-                      fill: "currentcolor",
-                    }}
-                  >
-                    {sortDirection === "ASC" && <path d="M0 8 6 0 12 8"></path>}
-                    {sortDirection === "DESC" && (
-                      <path d="M0 0 6 8 12 0"></path>
-                    )}
-                  </svg>
-                  {priority}
-                </span>
-              </span>
-            </div>
-            {filters.enabled && (
-              <div className={"filter-cell"}>
-                <input
-                  className="textInput"
-                  style={filterCSS}
-                  defaultValue={filters.columns[column.key]}
-                  onChange={handleInputKeyDown}
-                />
-              </div>
-            )}
-          </React.Fragment>
-        );
-      };
-    }
-
-    if (column.key === "pcntOfSession") {
-      addPercentage(column);
-    }
-  });
-
-  function filterColumn(columnKey: string, value: any) {
-    var tempFilters = filters;
-
-    if (value === "") {
-      delete tempFilters.columns[columnKey];
-    } else {
-      tempFilters.columns[columnKey] = value;
-    }
-
-    setFilters(tempFilters);
-
-    if (Object.keys(filters.columns).length === 0) {
-      setFilteredModuleRows(moduleRows);
-    } else {
-      setFilteredModuleRows(
-        moduleRows.filter((row) => {
-          for (let [key] of Object.entries(filters.columns)) {
-            if (
-              !row[key]
-                .toString()
-                .toLowerCase()
-                .includes(filters.columns[key].toLowerCase())
-            ) {
-              return false;
-            }
-          }
-          return true;
-        })
-      );
-    }
-  }
-
-  function addPercentage(column) {
-    column["headerRenderer"] = function ({}) {
-      return <span>{column.name}</span>;
-    };
-
-    column["formatter"] = function ({ row }) {
-      const percentage = row[column.key];
-      const progressStyle: React.CSSProperties = {
-        width: `${percentage}%`,
-        height: "10px",
-        backgroundColor: "#007bff",
-      };
-
-      const borderStyle: React.CSSProperties = {
-        border: "1px solid #ccc",
-        boxSizing: "border-box",
-      };
-
-      return (
-        <div className="progress" style={{ height: "10px" }}>
-          <div className="progress-border" style={borderStyle}>
-            <div
-              className="progress-bar"
-              role="progressbar"
-              style={progressStyle}
-              aria-valuenow={percentage}
-              aria-valuemin={0}
-              aria-valuemax={100}
-            >
-              {percentage}%
-            </div>
-          </div>
-        </div>
-      );
-    };
-  }
-
-  React.useLayoutEffect(() => {
-    window.addEventListener("message", (event) => {
-      const message = event.data as PresentationData;
-
-      setModuleRows(message.moduleDetails);
-      setFilteredModuleRows(message.moduleDetails);
-      setFilters({
-        columns: {},
-        enabled: true,
-      });
-
-      setCallingRows(message.calledModules);
-      setCalledRows(message.calledModules);
-      setLineRows(message.lineSummary);
-    });
-  }, []);
-
-  const showSelected = (row) => {
-    onRowSelect(row);
-    filterTables(row);
-  };
-
-  function filterTables(row) {
-    if (!row) return;
-    setSelectedCallingRows(
-      callingRows.filter((element) => element.calleeID === row.moduleID)
-    );
-    setSelectedCalledRows(
-      calledRows.filter((element) => element.callerID === row.moduleID)
-    );
-    setSelectedLineRows(
-      lineRows.filter((element) => element.moduleID === row.moduleID)
-    );
-  }
 
   React.useEffect(() => {
     filterTables(selectedRow);
   }, [selectedRow]);
+
+  React.useEffect(() => {
+    const { hasXREFs, hasListings } = presentationData;
+    if (hasXREFs && !hasListings) {
+      settingsContext.setOpenFileType(OpenFileTypeEnum.XREF);
+    } else if (!hasXREFs && hasListings) {
+      settingsContext.setOpenFileType(OpenFileTypeEnum.LISTING);
+    }
+  }, [presentationData.hasXREFs, presentationData.hasListings]);
 
   const openFileForLineSummary = (row: LineSummary): void => {
     if (!row.hasLink) {
@@ -470,113 +262,105 @@ function ProfilerModuleDetails({
       (moduleRow) => moduleRow.moduleID === row.moduleID
     );
 
-    const moduleName = moduleRow.moduleName;
-
-    vscode.postMessage({
-      type: "MODULE_NAME",
-      columns: moduleName,
-      lines: row.lineNumber,
-    });
-  };
-
-  const openFileForModuleDetails = (row: ModuleDetails): void => {
-    if (!row.hasLink) {
-      return;
+    switch (settingsContext.openFileType) {
+      case OpenFileTypeEnum.XREF:
+        vscode.postMessage({
+          type: OpenFileTypeEnum.XREF,
+          columns: moduleRow.moduleName,
+          lines: row.lineNumber,
+        });
+        break;
+      case OpenFileTypeEnum.LISTING:
+        vscode.postMessage({
+          type: OpenFileTypeEnum.LISTING,
+          listingFile: moduleRow.listingFile,
+          lineNumber: row.lineNumber,
+        });
+        break;
     }
-
-    vscode.postMessage({
-      type: "MODULE_NAME",
-      columns: row.moduleName,
-      lines: row.startLineNum,
-    });
   };
 
   return (
-    <React.Fragment>
-      <div>
-        <div className="details-columns">
-          <div className="grid-name">Module Details</div>
-          {moduleRows.length > 0 ? (
-            <DataGrid
-              columns={formattedModuleColumns}
-              // columns={columnDefinition.moduleColumns}
-              rows={sortedModuleRows}
-              defaultColumnOptions={{
-                sortable: true,
-                resizable: true,
-              }}
-              onRowClick={showSelected}
-              headerRowHeight={filters.enabled ? 70 : undefined}
-              rowKeyGetter={moduleRowKeyGetter}
-              onRowsChange={setModuleRows}
-              sortColumns={sortModuleColumns}
-              onSortColumnsChange={setSortModuleColumns}
-              onRowDoubleClick={openFileForModuleDetails}
-              rowClass={(row) => (row === selectedRow ? "rowFormat" : "")}
-            />
-          ) : null}
-          <div className="total-time">
-            Total Time: {sumTotalTime.toFixed(6)}s
-          </div>
-        </div>
-        <div className="columns">
-          <div className="calling-columns">
-            <div className="grid-name">Calling Modules</div>
-            <DataGrid
-              className="columns"
-              columns={callingColumns}
-              rows={sortedCallingRows}
-              defaultColumnOptions={{
-                sortable: true,
-                resizable: true,
-              }}
-              onRowsChange={setSelectedCallingRows}
-              sortColumns={sortCallingColumns}
-              onSortColumnsChange={setSortCallingColumns}
-              onRowDoubleClick={(row) => {
-                filterByModuleName(row.callerModuleName);
-              }}
-            />
-          </div>
-
-          <div className="called-columns">
-            <div className="grid-name">Called Modules</div>
-            <DataGrid
-              className="columns"
-              columns={calledColumns}
-              rows={sortedCalledRows}
-              defaultColumnOptions={{
-                sortable: true,
-                resizable: true,
-              }}
-              onRowsChange={setSelectedCalledRows}
-              sortColumns={sortCalledColumns}
-              onSortColumnsChange={setSortCalledColumns}
-              onRowDoubleClick={(row) => {
-                filterByModuleName(row.calleeModuleName);
-              }}
-            />
-          </div>
-        </div>
-
-        <div className="line-columns">
-          <div className="grid-name">Line Summary</div>
+    <div>
+      <div className="details-columns">
+        <div className="grid-name">Module Details</div>
+        <ModuleDetailsSettings
+          showOpenFileType={
+            presentationData.hasXREFs && presentationData.hasListings
+          }
+        />
+        {moduleRows.length > 0 ? (
+          <ModuleDetailsTable
+            columns={formattedModuleColumns}
+            rows={sortedModuleRows}
+            onRowClick={(row) => setSelectedRow(row)}
+            onRowsChange={setModuleRows}
+            sortColumns={sortModuleColumns}
+            onSortColumnsChange={setSortModuleColumns}
+            rowClass={(row) => (row === selectedRow ? "rowFormat" : "")}
+            sumTotalTime={sumTotalTime}
+            searchValue={moduleNameFilter}
+            setSearchValue={setModuleNameFilter}
+          />
+        ) : null}
+      </div>
+      <div className="columns">
+        <div className="calling-columns">
+          <div className="grid-name">Calling Modules</div>
           <DataGrid
-            columns={formattedLineColumns}
-            rows={sortedLineRows}
+            className="columns"
+            columns={callingColumns}
+            rows={sortedCallingRows}
             defaultColumnOptions={{
               sortable: true,
               resizable: true,
             }}
-            onRowsChange={setSelectedLineRows}
-            sortColumns={sortLineColumns}
-            onSortColumnsChange={setSortLineColumns}
-            onRowDoubleClick={openFileForLineSummary}
+            onRowsChange={setSelectedCallingRows}
+            sortColumns={sortCallingColumns}
+            onSortColumnsChange={setSortCallingColumns}
+            onRowDoubleClick={(row) => {
+              setModuleNameFilter(row.callerModuleName);
+            }}
+          />
+        </div>
+
+        <div className="called-columns">
+          <div className="grid-name">Called Modules</div>
+          <DataGrid
+            className="columns"
+            columns={calledColumns}
+            rows={sortedCalledRows}
+            defaultColumnOptions={{
+              sortable: true,
+              resizable: true,
+            }}
+            onRowsChange={setSelectedCalledRows}
+            sortColumns={sortCalledColumns}
+            onSortColumnsChange={setSortCalledColumns}
+            onRowDoubleClick={(row) => {
+              setModuleNameFilter(row.calleeModuleName);
+            }}
           />
         </div>
       </div>
-    </React.Fragment>
+
+      <div className="line-columns">
+        <div className="grid-name">Line Summary</div>
+        <DataGrid
+          columns={formattedLineColumns}
+          rows={sortedLineRows}
+          defaultColumnOptions={{
+            sortable: true,
+            resizable: true,
+          }}
+          onRowsChange={setSelectedLineRows}
+          sortColumns={sortLineColumns}
+          onSortColumnsChange={setSortLineColumns}
+          onRowDoubleClick={openFileForLineSummary}
+        />
+      </div>
+    </div>
   );
-}
+};
 
 export default ProfilerModuleDetails;