+++ src/view/app/Profiler/profilerForm.tsx
@@ -4,28 +4,39 @@ import { PresentationData } from "../../../common/PresentationData";
 import ProfilerTreeView from "../ProfilerTreeView/profilerTreeView";
 import ProfilerFlameGraph from "../FlameGraph/profilerFlameGraph";
 import ProfilerModuleDetails from "../ModuleDetails/profilerModuleDetails";
-import { Button } from "@mui/material";
+import { ToggleButtonGroup } from "@mui/material";
 import LoadingOverlay from "../../../../src/components/loadingOverlay/loadingOverlay";
+import { getVSCodeAPI } from "../utils/vscode";
+import ModuleDetailsSettingsContextProvider from "../ModuleDetails/components/ModuleDetailsSettingsContext";
+import ProToggleButton from "../Components/Buttons/ProToggleButton";
 
-interface IConfigProps {
-  vscode: any;
-  presentationData: PresentationData;
-}
+const defaultPresentationData: PresentationData = {
+  moduleDetails: [],
+  calledModules: [],
+  callTree: [],
+  lineSummary: [],
+  hasTracingData: false,
+  hasXREFs: false,
+  hasListings: false,
+};
 
 enum ProfilerTab {
-  ModuleDetails = 0,
-  TreeView = 1,
-  FlameGraph = 2,
+  ModuleDetails = "ModuleDetails",
+  TreeView = "TreeView",
+  FlameGraph = "FlameGraph",
 }
 
-function ProfilerForm({ presentationData, vscode }: IConfigProps) {
+const ProfilerForm: React.FC = () => {
   const [activeTab, setActiveTab] = useState<ProfilerTab>(
     ProfilerTab.ModuleDetails
   );
-  const [presentationData2, setPresentationData] = useState(presentationData);
+  const [presentationData, setPresentationData] = useState<PresentationData>(
+    defaultPresentationData
+  );
   const [isLoading, setLoading] = useState(true);
-  const [selectedRow, setSelectedRow] = useState<any>(null);
   const [moduleName, setModuleName] = useState<string>("");
+
+  const vscode = getVSCodeAPI();
 
   React.useLayoutEffect(() => {
     window.addEventListener("message", (event) => {
@@ -39,13 +50,12 @@ function ProfilerForm({ presentationData, vscode }: IConfigProps) {
   const ModuleDetailsTab: React.FC = () => {
     return (
       <div>
-        <ProfilerModuleDetails
-          presentationData={presentationData2}
-          vscode={vscode}
-          selectedRow={selectedRow}
-          onRowSelect={handleRowSelection}
-          moduleName={moduleName}
-        />
+        <ModuleDetailsSettingsContextProvider>
+          <ProfilerModuleDetails
+            presentationData={presentationData}
+            moduleName={moduleName}
+          />
+        </ModuleDetailsSettingsContextProvider>
       </div>
     );
   };
@@ -54,7 +64,7 @@ function ProfilerForm({ presentationData, vscode }: IConfigProps) {
     return (
       <div>
         <ProfilerTreeView
-          presentationData={presentationData2}
+          presentationData={presentationData}
           handleNodeSelection={handleNodeSelection}
         />
       </div>
@@ -65,8 +75,8 @@ function ProfilerForm({ presentationData, vscode }: IConfigProps) {
     return (
       <div>
         <ProfilerFlameGraph
-          presentationData={presentationData2}
-          hasTracingData={presentationData2.hasTracingData}
+          presentationData={presentationData}
+          hasTracingData={presentationData.hasTracingData}
           handleNodeSelection={handleNodeSelection}
           vscode={vscode}
         />
@@ -76,16 +86,17 @@ function ProfilerForm({ presentationData, vscode }: IConfigProps) {
 
   let content: JSX.Element | null = null;
 
-  const handleTabClick = (tab: ProfilerTab) => {
-    setActiveTab(tab);
+  const onTabChange = (
+    event: React.MouseEvent<HTMLElement>,
+    tab: ProfilerTab | null
+  ) => {
+    console.log(tab);
 
-    if (activeTab !== ProfilerTab.ModuleDetails) {
-      setModuleName("");
+    if (!tab) {
+      return;
     }
-  };
 
-  const handleRowSelection = (row: any) => {
-    setSelectedRow(row);
+    setActiveTab(tab);
   };
 
   const handleNodeSelection = (moduleName: string) => {
@@ -93,17 +104,15 @@ function ProfilerForm({ presentationData, vscode }: IConfigProps) {
     setActiveTab(ProfilerTab.ModuleDetails);
   };
 
+  React.useEffect(() => {
+    if (activeTab !== ProfilerTab.ModuleDetails) {
+      setModuleName("");
+    }
+  }, [activeTab]);
+
   switch (activeTab) {
     case ProfilerTab.ModuleDetails:
-      content = (
-        <ProfilerModuleDetails
-          presentationData={presentationData2}
-          vscode={vscode}
-          selectedRow={selectedRow}
-          onRowSelect={handleRowSelection}
-          moduleName={moduleName}
-        />
-      );
+      content = <ModuleDetailsTab />;
       break;
     case ProfilerTab.TreeView:
       content = <TreeViewTab />;
@@ -117,38 +126,24 @@ function ProfilerForm({ presentationData, vscode }: IConfigProps) {
     <React.Fragment>
       {isLoading && <LoadingOverlay />}
       <div>
-        <div className="tabs">
-          <Button
-            className={`tab ${
-              activeTab === ProfilerTab.ModuleDetails ? "active" : ""
-            } buttonProfilerForm button-primary`}
-            onClick={() => handleTabClick(ProfilerTab.ModuleDetails)}
-            variant="contained"
-          >
-            Module Details
-          </Button>
-          <Button
-            className={`tab ${
-              activeTab === ProfilerTab.TreeView ? "active" : ""
-            } buttonProfilerForm button-primary`}
-            onClick={() => handleTabClick(ProfilerTab.TreeView)}
-          >
-            Tree View
-          </Button>
-          <Button
-            className={`tab ${
-              activeTab === ProfilerTab.FlameGraph ? "active" : ""
-            } buttonProfilerForm button-primary`}
-            onClick={() => handleTabClick(ProfilerTab.FlameGraph)}
-          >
-            Flame Graph
-          </Button>
+        <div>
+          <ToggleButtonGroup value={activeTab} onChange={onTabChange} exclusive>
+            <ProToggleButton value={ProfilerTab.ModuleDetails}>
+              Module Details
+            </ProToggleButton>
+            <ProToggleButton value={ProfilerTab.TreeView}>
+              Tree View
+            </ProToggleButton>
+            <ProToggleButton value={ProfilerTab.FlameGraph}>
+              Flame Graph
+            </ProToggleButton>
+          </ToggleButtonGroup>
         </div>
         <hr></hr>
         <div>{content}</div>
       </div>
     </React.Fragment>
   );
-}
+};
 
 export default ProfilerForm;